[fiber] Reorder includes to isolate common function
