+++ src/modm/processing/fiber/condition_variable.hpp
@@ -11,7 +11,7 @@
 
 #pragma once
 
-#include "fiber.hpp"
+#include "functions.hpp"
 #include "stop_token.hpp"
 #include <atomic>
