+++ src/modm/processing/fiber/semaphore.hpp
@@ -11,7 +11,7 @@
 
 #pragma once
 
-#include "fiber.hpp"
+#include "functions.hpp"
 #include <limits>
 #include <atomic>
