+++ src/modm/processing/fiber/shared_mutex.hpp
@@ -11,7 +11,7 @@
 
 #pragma once
 
-#include "fiber.hpp"
+#include "functions.hpp"
 #include <atomic>
 #include <shared_mutex>
